requirements.txt
python-dotenv~=1.0.1
openai~=1.52.1
pydantic~=2.8.2
selenium~=4.27.1
schedule~=1.2.2
attrs~=24.2.0
beautifulsoup4~=4.12.3
pyspark~=3.4.1
hdfs~=2.7.3
flask~=3.0.3
pandas~=2.2.3
clickhouse-driver~=0.2.9
numpy~=1.26.4
scikit-learn~=1.6.0
pytz~=2024.1
setuptools~=75.6.0
cachetools~=5.5.0
flask-cors~=5.0.0